add maze line part

+++ src/maps_pogram/maze_line/maze_line.h
@@ -6,4 +6,4 @@ void mazeLineSetup();
 
 void mazeLine(int speed);
 
-float getPWMValue(int speed);
+int getPWMValue(int speed);
